+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 11.8.2
 -----------------
 - CSS fix input div onclick & .underline-dots line height in stylesheet.css
+- Fix .thu-fri-sat class for Calendar day on Public Pages in CalendarDay.inc.php
 
 Changes in 11.8.1
 -----------------